refactor(kafka): update ResourceEvent imports and adjust Kafka utility references
