+++ NEWS.md
@@ -16,6 +16,7 @@
 * Half of timers are disabled after recent bugfest update for 1 tenant (MGRENTITLE-165)
 * Migrate CI to centralized FOLIO Maven GitHub workflow (MGRENTITLE-162)
 * Upgrade module to SpringBoot4.0 and Spring7.0 (MGRENTITLE-148)
+* Adopt APPPOCTOOL-85 Kafka producer/common module split and shared resource event types (APPPOCTOOL-85)
 
 ---
