+++ pom.xml
@@ -37,7 +37,7 @@
     <cql2pgjson.version>36.0.0</cql2pgjson.version>
     <openapi-tools.jackson-databind-nullable.version>0.2.10</openapi-tools.jackson-databind-nullable.version>
     <folio-spring-cql.version>10.0.0</folio-spring-cql.version>
-    <applications-poc-tools.version>4.0.0-SNAPSHOT</applications-poc-tools.version>
+    <applications-poc-tools.version>4.1.0-SNAPSHOT</applications-poc-tools.version>
     <folio-java-checkstyle.version>1.2.0</folio-java-checkstyle.version>
     <folio-flow-engine.version>4.0.0-SNAPSHOT</folio-flow-engine.version>
     <caffeine.version>3.1.8</caffeine.version>
@@ -283,7 +283,13 @@
 
     <dependency>
       <groupId>org.folio</groupId>
-      <artifactId>folio-integration-kafka</artifactId>
+      <artifactId>folio-kafka-producer</artifactId>
+      <version>${applications-poc-tools.version}</version>
+    </dependency>
+
+    <dependency>
+      <groupId>org.folio</groupId>
+      <artifactId>folio-kafka-common</artifactId>
       <version>${applications-poc-tools.version}</version>
     </dependency>
