+++ src/main/java/org/folio/entitlement/integration/kafka/EntitlementEventPublisher.java
@@ -1,6 +1,6 @@
 package org.folio.entitlement.integration.kafka;
 
-import static org.folio.integration.kafka.KafkaUtils.getEnvTopicName;
+import static org.folio.integration.kafka.producer.KafkaUtils.getEnvTopicName;
 
 import lombok.RequiredArgsConstructor;
 import org.folio.entitlement.integration.kafka.model.EntitlementEvent;
