+++ src/main/java/org/folio/entitlement/integration/kafka/ScheduledJobEventPublisher.java
@@ -3,7 +3,7 @@ package org.folio.entitlement.integration.kafka;
 import static org.folio.entitlement.integration.kafka.KafkaEventUtils.SCHEDULED_JOB_RESOURCE_NAME;
 import static org.folio.entitlement.integration.kafka.KafkaEventUtils.SCHEDULED_JOB_TOPIC;
 import static org.folio.entitlement.integration.kafka.ScheduledJobModuleEventPublisher.getScheduledTimers;
-import static org.folio.integration.kafka.KafkaUtils.getTenantTopicName;
+import static org.folio.integration.kafka.producer.KafkaUtils.getTenantTopicName;
 
 import java.util.Optional;
 import org.folio.common.domain.model.ModuleDescriptor;
