+++ src/main/java/org/folio/entitlement/integration/kafka/SystemUserModuleEventPublisher.java
@@ -3,7 +3,7 @@ package org.folio.entitlement.integration.kafka;
 import static org.folio.entitlement.integration.kafka.KafkaEventUtils.SYSTEM_USER_RESOURCE_NAME;
 import static org.folio.entitlement.integration.kafka.KafkaEventUtils.SYSTEM_USER_TOPIC;
 import static org.folio.entitlement.integration.kafka.KafkaEventUtils.TOPIC_TENANT_COLLECTION_KEY;
-import static org.folio.integration.kafka.KafkaUtils.getTenantTopicName;
+import static org.folio.integration.kafka.producer.KafkaUtils.getTenantTopicName;
 
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
