+++ src/main/java/org/folio/entitlement/integration/kafka/configuration/TenantEntitlementKafkaProperties.java
@@ -3,7 +3,7 @@ package org.folio.entitlement.integration.kafka.configuration;
 import java.time.Duration;
 import java.util.List;
 import lombok.Data;
-import org.folio.integration.kafka.FolioKafkaProperties.KafkaTopic;
+import org.folio.integration.kafka.producer.KafkaProducerProperties.KafkaTopic;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.NestedConfigurationProperty;
 import org.springframework.stereotype.Component;
