--- src/main/java/org/folio/entitlement/integration/kafka/model/ResourceEventType.java
@@ -1,18 +1,0 @@
-package org.folio.entitlement.integration.kafka.model;
-
-import com.fasterxml.jackson.annotation.JsonValue;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum ResourceEventType {
-
-  UPDATE("UPDATE"),
-  CREATE("CREATE"),
-  DELETE("DELETE"),
-  DELETE_ALL("DELETE_ALL");
-
-  @JsonValue
-  private final String value;
-}
