+++ src/test/java/org/folio/entitlement/integration/kafka/CapabilitiesEventPublisherTest.java
@@ -40,8 +40,8 @@ import java.util.stream.Stream;
 import org.folio.common.domain.model.ApplicationDescriptor;
 import org.folio.entitlement.domain.model.EntitlementRequest;
 import org.folio.entitlement.integration.kafka.model.CapabilityEventPayload;
-import org.folio.entitlement.integration.kafka.model.ResourceEvent;
 import org.folio.entitlement.support.TestUtils;
+import org.folio.integration.kafka.model.ResourceEvent;
 import org.folio.test.types.UnitTest;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
