+++ src/test/java/org/folio/entitlement/integration/kafka/ScheduledJobModuleEventPublisherTest.java
@@ -12,9 +12,6 @@ import static org.folio.entitlement.domain.model.CommonStageContext.PARAM_REQUES
 import static org.folio.entitlement.domain.model.CommonStageContext.PARAM_TENANT_NAME;
 import static org.folio.entitlement.domain.model.ModuleStageContext.PARAM_INSTALLED_MODULE_DESCRIPTOR;
 import static org.folio.entitlement.domain.model.ModuleStageContext.PARAM_MODULE_DESCRIPTOR;
-import static org.folio.entitlement.integration.kafka.model.ResourceEventType.CREATE;
-import static org.folio.entitlement.integration.kafka.model.ResourceEventType.DELETE;
-import static org.folio.entitlement.integration.kafka.model.ResourceEventType.UPDATE;
 import static org.folio.entitlement.support.TestConstants.APPLICATION_FLOW_ID;
 import static org.folio.entitlement.support.TestConstants.APPLICATION_ID;
 import static org.folio.entitlement.support.TestConstants.ENTITLED_APPLICATION_ID;
@@ -26,6 +23,9 @@ import static org.folio.entitlement.support.TestConstants.scheduledJobsTenantCol
 import static org.folio.entitlement.support.TestConstants.scheduledJobsTenantTopic;
 import static org.folio.entitlement.support.TestValues.moduleFlowParameters;
 import static org.folio.entitlement.support.TestValues.moduleStageContext;
+import static org.folio.integration.kafka.model.ResourceEventType.CREATE;
+import static org.folio.integration.kafka.model.ResourceEventType.DELETE;
+import static org.folio.integration.kafka.model.ResourceEventType.UPDATE;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
@@ -37,9 +37,9 @@ import org.folio.common.domain.model.ModuleDescriptor;
 import org.folio.common.domain.model.RoutingEntry;
 import org.folio.entitlement.domain.model.EntitlementRequest;
 import org.folio.entitlement.integration.kafka.configuration.TenantEntitlementKafkaProperties;
-import org.folio.entitlement.integration.kafka.model.ResourceEvent;
 import org.folio.entitlement.integration.kafka.model.ScheduledTimers;
 import org.folio.entitlement.support.TestUtils;
+import org.folio.integration.kafka.model.ResourceEvent;
 import org.folio.test.types.UnitTest;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
