+++ src/test/java/org/folio/entitlement/it/AbstractFolioEntitlementIT.java
@@ -66,10 +66,10 @@ import org.awaitility.Awaitility;
 import org.folio.common.utils.OkapiHeaders;
 import org.folio.entitlement.domain.dto.DesiredStateRequestBody;
 import org.folio.entitlement.integration.kafka.model.EntitlementEvent;
-import org.folio.entitlement.integration.kafka.model.ResourceEvent;
 import org.folio.entitlement.support.KeycloakTestClientConfiguration;
 import org.folio.entitlement.support.KeycloakTestClientConfiguration.KeycloakTestClient;
 import org.folio.entitlement.support.base.BaseIntegrationTest;
+import org.folio.integration.kafka.model.ResourceEvent;
 import org.folio.test.extensions.EnableKeycloakTlsMode;
 import org.folio.test.extensions.KeycloakRealms;
 import org.folio.test.extensions.WireMockStub;
