+++ src/test/java/org/folio/entitlement/it/NoIntegrationsFolioEntitlementIT.java
@@ -62,8 +62,8 @@ import java.util.Map;
 import org.apache.commons.lang3.tuple.Pair;
 import org.folio.entitlement.domain.dto.FlowStage;
 import org.folio.entitlement.integration.kafka.model.EntitlementEvent;
-import org.folio.entitlement.integration.kafka.model.ResourceEvent;
 import org.folio.entitlement.support.base.BaseIntegrationTest;
+import org.folio.integration.kafka.model.ResourceEvent;
 import org.folio.test.extensions.EnableKeycloakTlsMode;
 import org.folio.test.extensions.KeycloakRealms;
 import org.folio.test.extensions.WireMockStub;
