+++ src/test/java/org/folio/entitlement/support/DesiredStateTestValues.java
@@ -16,8 +16,8 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.folio.entitlement.integration.kafka.model.CapabilityEventPayload;
 import org.folio.entitlement.integration.kafka.model.EntitlementEvent;
-import org.folio.entitlement.integration.kafka.model.ResourceEvent;
 import org.folio.entitlement.support.model.AuthorizationResource;
+import org.folio.integration.kafka.model.ResourceEvent;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class DesiredStateTestValues {
