+++ src/test/java/org/folio/entitlement/support/TestUtils.java
@@ -37,12 +37,12 @@ import org.folio.common.domain.model.ApplicationDescriptor;
 import org.folio.common.domain.model.ModuleDescriptor;
 import org.folio.entitlement.exception.RequestValidationException;
 import org.folio.entitlement.integration.kafka.model.CapabilityEventPayload;
-import org.folio.entitlement.integration.kafka.model.ResourceEvent;
 import org.folio.entitlement.integration.kafka.model.ScheduledTimers;
 import org.folio.entitlement.integration.kafka.model.SystemUserEvent;
 import org.folio.flow.api.Stage;
 import org.folio.flow.exception.StageExecutionException;
 import org.folio.flow.model.StageResult;
+import org.folio.integration.kafka.model.ResourceEvent;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.Spy;
